pending query cache design notes updated and interface defined and documented
